QQuickThreadedRenderLoop: clear polish timer before flushing events
When calling QWindowSystemInterface::flushWindowSystemEvents() from
QSGThreadedRenderLoop::polishAndSync(), new touch event might be
delivered to QtQuickWindow. This can cause
QQuickWindowPrivate::deliverTouchEvent() to call
QSGThreadedRenderLoop::maybeUpdate(), which tries to start a new
polish timer. The problem is that this timer will be killed
later down in polishAndSync, and hence, the update will be forgotten.

A problem with this can sometimes be seen when dragging on a
MouseArea. If we fail to polish after receiving a touch event,
the touch compression algorithm will just continue to
compress subsequent events waiting for the flush that never comes.
This again will freeze the whole drag.

This patch will move killing of the timer before we flush window
system events so that a possible update call will restart the timer.

Change-Id: I897b9329b39fc3a842271d98c2df9e0ae58eec3a